+++ R/linear_reg.R
@@ -11,7 +11,7 @@
 #'   regularization (see below). Note that this will be ignored for some engines.
 #' }
 #' These arguments are converted to their specific names at the
-#'  time that the model is fit. Other options and argument can be
+#'  time that the model is fit. Other options and arguments can be
 #'  set using `set_engine()`. If left to their defaults
 #'  here (`NULL`), the values are taken from the underlying model
 #'  functions. If parameters need to be modified, `update()` can be used
@@ -42,6 +42,9 @@
 #' \item \pkg{keras}: `"keras"`
 #' }
 #'
+#' For this model, other packages may add additional engines. Use
+#' [show_engines()] to see the current set of engines.
+#'
 #' @includeRmd man/rmd/linear-reg.Rmd details
 #'
 #' @note For models created using the spark engine, there are
@@ -59,6 +62,8 @@
 #'
 #' @seealso [fit()], [set_engine()]
 #' @examples
+#' show_engines("linear_reg")
+#'
 #' linear_reg()
 #' # Parameters can be represented by a placeholder:
 #' linear_reg(penalty = varying())