+++ R/logistic_reg_data.R
@@ -26,7 +26,8 @@ set_encoding(
   options = list(
     predictor_indicators = "traditional",
     compute_intercept = TRUE,
-    remove_intercept = TRUE
+    remove_intercept = TRUE,
+    allow_sparse_x = FALSE
   )
 )
 
@@ -151,7 +152,8 @@ set_encoding(
   options = list(
     predictor_indicators = "traditional",
     compute_intercept = TRUE,
-    remove_intercept = TRUE
+    remove_intercept = TRUE,
+    allow_sparse_x = TRUE
   )
 )
 
@@ -184,10 +186,10 @@ set_pred(
     func = c(fun = "predict"),
     args =
       list(
-        object = quote(object$fit),
-        newx = quote(as.matrix(new_data)),
+        object = expr(object$fit),
+        newx = expr(as.matrix(new_data[, rownames(object$fit$beta)])),
         type = "response",
-        s = quote(object$spec$args$penalty)
+        s = expr(object$spec$args$penalty)
       )
   )
 )
@@ -274,7 +276,8 @@ set_encoding(
   options = list(
     predictor_indicators = "traditional",
     compute_intercept = TRUE,
-    remove_intercept = TRUE
+    remove_intercept = TRUE,
+    allow_sparse_x = FALSE
   )
 )
 
@@ -346,7 +349,8 @@ set_encoding(
   options = list(
     predictor_indicators = "traditional",
     compute_intercept = TRUE,
-    remove_intercept = TRUE
+    remove_intercept = TRUE,
+    allow_sparse_x = FALSE
   )
 )
 
@@ -414,7 +418,8 @@ set_encoding(
   options = list(
     predictor_indicators = "traditional",
     compute_intercept = TRUE,
-    remove_intercept = TRUE
+    remove_intercept = TRUE,
+    allow_sparse_x = FALSE
   )
 )