+++ R/surv_reg.R
@@ -8,7 +8,7 @@
 #'   \item \code{dist}: The probability distribution of the outcome.
 #' }
 #' This argument is converted to its specific names at the
-#'  time that the model is fit. Other options and argument can be
+#'  time that the model is fit. Other options and arguments can be
 #'  set using `set_engine()`. If left to its default
 #'  here (`NULL`), the value is taken from the underlying model
 #'  functions.
@@ -45,6 +45,8 @@
 #' @references Jackson, C. (2016). `flexsurv`: A Platform for Parametric Survival
 #'  Modeling in R. _Journal of Statistical Software_, 70(8), 1 - 33.
 #' @examples
+#' show_engines("surv_reg")
+#'
 #' surv_reg()
 #' # Parameters can be represented by a placeholder:
 #' surv_reg(dist = varying())