+++ tests/testthat/test_boost_tree_xgboost.R
@@ -122,6 +122,16 @@ test_that('xgboost execution, regression', {
     ),
     regexp = NA
   )
+
+  expect_error(
+    res <- parsnip::fit_xy(
+      car_basic,
+      x = mtcars[, num_pred],
+      y = factor(mtcars$vs),
+      control = ctrl
+    ),
+    regexp = "For a regression model"
+  )
 })
 
 
@@ -286,3 +296,97 @@ test_that('early stopping', {
    regex = "`early_stop` should be on"
  )
 })
+
+
+## -----------------------------------------------------------------------------
+
+test_that('xgboost data conversion', {
+  skip_if_not_installed("xgboost")
+
+  mtcar_x <- mtcars[, -1]
+  mtcar_mat <- as.matrix(mtcar_x)
+  mtcar_smat <- Matrix::Matrix(mtcar_mat, sparse = TRUE)
+
+  expect_error(from_df <- parsnip:::as_xgb_data(mtcar_x, mtcars$mpg), regexp = NA)
+  expect_true(inherits(from_df$data, "xgb.DMatrix"))
+  expect_true(inherits(from_df$watchlist$training, "xgb.DMatrix"))
+
+  expect_error(from_mat <- parsnip:::as_xgb_data(mtcar_mat, mtcars$mpg), regexp = NA)
+  expect_true(inherits(from_mat$data, "xgb.DMatrix"))
+  expect_true(inherits(from_mat$watchlist$training, "xgb.DMatrix"))
+
+  expect_error(from_sparse <- parsnip:::as_xgb_data(mtcar_smat, mtcars$mpg), regexp = NA)
+  expect_true(inherits(from_mat$data, "xgb.DMatrix"))
+  expect_true(inherits(from_mat$watchlist$training, "xgb.DMatrix"))
+
+  expect_error(from_df <- parsnip:::as_xgb_data(mtcar_x, mtcars$mpg, validation = .1), regexp = NA)
+  expect_true(inherits(from_df$data, "xgb.DMatrix"))
+  expect_true(inherits(from_df$watchlist$validation, "xgb.DMatrix"))
+  expect_true(nrow(from_df$data) > nrow(from_df$watchlist$validation))
+
+  expect_error(from_mat <- parsnip:::as_xgb_data(mtcar_mat, mtcars$mpg, validation = .1), regexp = NA)
+  expect_true(inherits(from_mat$data, "xgb.DMatrix"))
+  expect_true(inherits(from_mat$watchlist$validation, "xgb.DMatrix"))
+  expect_true(nrow(from_mat$data) > nrow(from_mat$watchlist$validation))
+
+  expect_error(from_sparse <- parsnip:::as_xgb_data(mtcar_smat, mtcars$mpg, validation = .1), regexp = NA)
+  expect_true(inherits(from_mat$data, "xgb.DMatrix"))
+  expect_true(inherits(from_mat$watchlist$validation, "xgb.DMatrix"))
+  expect_true(nrow(from_sparse$data) > nrow(from_sparse$watchlist$validation))
+
+})
+
+
+test_that('xgboost data and sparse matrices', {
+  skip_if_not_installed("xgboost")
+
+  mtcar_x <- mtcars[, -1]
+  mtcar_mat <- as.matrix(mtcar_x)
+  mtcar_smat <- Matrix::Matrix(mtcar_mat, sparse = TRUE)
+
+  xgb_spec <-
+    boost_tree(trees = 10) %>%
+    set_engine("xgboost") %>%
+    set_mode("regression")
+
+  set.seed(1)
+  from_df <- xgb_spec %>% fit_xy(mtcar_x, mtcars$mpg)
+  set.seed(1)
+  from_mat <- xgb_spec %>% fit_xy(mtcar_mat, mtcars$mpg)
+  set.seed(1)
+  from_sparse <- xgb_spec %>% fit_xy(mtcar_smat, mtcars$mpg)
+
+  expect_equal(from_df$fit, from_mat$fit)
+  expect_equal(from_df$fit, from_sparse$fit)
+
+})
+
+
+## -----------------------------------------------------------------------------
+
+test_that('argument checks for data dimensions', {
+
+  skip_if_not_installed("C50")
+
+  data(penguins, package = "modeldata")
+  penguins <- na.omit(penguins)
+
+  spec <-
+    boost_tree(mtry = 1000, min_n = 1000, trees = 5) %>%
+    set_engine("xgboost") %>%
+    set_mode("classification")
+
+  penguins_dummy <- model.matrix(species ~ ., data = penguins)
+  penguins_dummy <- as.data.frame(penguins_dummy[, -1])
+
+  f_fit  <- spec %>% fit(species ~ ., data = penguins)
+  xy_fit <- spec %>% fit_xy(x = penguins_dummy, y = penguins$species)
+
+  expect_equal(f_fit$fit$params$colsample_bytree, 1)
+  expect_equal(f_fit$fit$params$min_child_weight, nrow(penguins))
+  expect_equal(xy_fit$fit$params$colsample_bytree, 1)
+  expect_equal(xy_fit$fit$params$min_child_weight, nrow(penguins))
+
+})
+
+