+++ tests/testthat/test_rand_forest.R
@@ -16,10 +16,10 @@ test_that('primary arguments', {
   mtry_spark <- translate(mtry %>% set_engine("spark"))
   expect_equal(mtry_ranger$method$fit$args,
                list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
+                 x = expr(missing_arg()),
+                 y = expr(missing_arg()),
                  case.weights = expr(missing_arg()),
-                 mtry = new_empty_quosure(4),
+                 mtry = expr(min_cols(~4, x)),
                  num.threads = 1,
                  verbose = FALSE,
                  seed = expr(sample.int(10^5, 1))
@@ -29,7 +29,7 @@ test_that('primary arguments', {
                list(
                  x = expr(missing_arg()),
                  y = expr(missing_arg()),
-                 mtry = new_empty_quosure(4)
+                 mtry = expr(min_cols(~4, x))
                )
   )
   expect_equal(mtry_spark$method$fit$args,
@@ -47,8 +47,8 @@ test_that('primary arguments', {
   trees_spark <- translate(trees %>% set_engine("spark"))
   expect_equal(trees_ranger$method$fit$args,
                list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
+                 x = expr(missing_arg()),
+                 y = expr(missing_arg()),
                  case.weights = expr(missing_arg()),
                  num.trees = new_empty_quosure(1000),
                  num.threads = 1,
@@ -80,10 +80,10 @@ test_that('primary arguments', {
   min_n_spark <- translate(min_n %>% set_engine("spark"))
   expect_equal(min_n_ranger$method$fit$args,
                list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
+                 x = expr(missing_arg()),
+                 y = expr(missing_arg()),
                  case.weights = expr(missing_arg()),
-                 min.node.size = new_empty_quosure(5),
+                 min.node.size = expr(min_rows(~5, x)),
                  num.threads = 1,
                  verbose = FALSE,
                  seed = expr(sample.int(10^5, 1))
@@ -93,7 +93,7 @@ test_that('primary arguments', {
                list(
                  x = expr(missing_arg()),
                  y = expr(missing_arg()),
-                 nodesize = new_empty_quosure(5)
+                 nodesize = expr(min_rows(~5, x))
                )
   )
   expect_equal(min_n_spark$method$fit$args,
@@ -101,117 +101,18 @@ test_that('primary arguments', {
                  x = expr(missing_arg()),
                  formula = expr(missing_arg()),
                  type = "regression",
-                 min_instances_per_node = new_empty_quosure(5),
+                 min_instances_per_node = expr(min_rows(~5, x)),
                  seed = expr(sample.int(10^5, 1))
                )
   )
-
-  mtry_v <- rand_forest(mode = "classification", mtry = varying())
-  mtry_v_ranger <- translate(mtry_v %>% set_engine("ranger"))
-  mtry_v_randomForest <- translate(mtry_v %>% set_engine("randomForest"))
-  mtry_v_spark <- translate(mtry_v %>% set_engine("spark"))
-  expect_equal(mtry_v_ranger$method$fit$args,
-               list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
-                 case.weights = expr(missing_arg()),
-                 mtry = new_empty_quosure(varying()),
-                 num.threads = 1,
-                 verbose = FALSE,
-                 seed = expr(sample.int(10^5, 1)),
-                 probability = TRUE
-               )
-  )
-  expect_equal(mtry_v_randomForest$method$fit$args,
-               list(
-                 x = expr(missing_arg()),
-                 y = expr(missing_arg()),
-                 mtry = new_empty_quosure(varying())
-               )
-  )
-  expect_equal(mtry_v_spark$method$fit$args,
-               list(
-                 x = expr(missing_arg()),
-                 formula = expr(missing_arg()),
-                 type = "classification",
-                 feature_subset_strategy = new_empty_quosure(varying()),
-                 seed = expr(sample.int(10^5, 1))
-               )
-  )
-
-  trees_v <- rand_forest(mode = "regression", trees = varying())
-  trees_v_ranger <- translate(trees_v %>% set_engine("ranger"))
-  trees_v_randomForest <- translate(trees_v %>% set_engine("randomForest"))
-  trees_v_spark <- translate(trees_v %>% set_engine("spark"))
-  expect_equal(trees_v_ranger$method$fit$args,
-               list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
-                 case.weights = expr(missing_arg()),
-                 num.trees = new_empty_quosure(varying()),
-                 num.threads = 1,
-                 verbose = FALSE,
-                 seed = expr(sample.int(10^5, 1))
-               )
-  )
-  expect_equal(trees_v_randomForest$method$fit$args,
-               list(
-                 x = expr(missing_arg()),
-                 y = expr(missing_arg()),
-                 ntree = new_empty_quosure(varying())
-               )
-  )
-  expect_equal(trees_v_spark$method$fit$args,
-               list(
-                 x = expr(missing_arg()),
-                 formula = expr(missing_arg()),
-                 type = "regression",
-                 num_trees = new_empty_quosure(varying()),
-                 seed = expr(sample.int(10^5, 1))
-               )
-  )
-
-  min_n_v <- rand_forest(mode = "classification", min_n = varying())
-  min_n_v_ranger <- translate(min_n_v %>% set_engine("ranger"))
-  min_n_v_randomForest <- translate(min_n_v %>% set_engine("randomForest"))
-  min_n_v_spark <- translate(min_n_v %>% set_engine("spark"))
-  expect_equal(min_n_v_ranger$method$fit$args,
-               list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
-                 case.weights = expr(missing_arg()),
-                 min.node.size = new_empty_quosure(varying()),
-                 num.threads = 1,
-                 verbose = FALSE,
-                 seed = expr(sample.int(10^5, 1)),
-                 probability = TRUE
-               )
-  )
-  expect_equal(min_n_v_randomForest$method$fit$args,
-               list(
-                 x = expr(missing_arg()),
-                 y = expr(missing_arg()),
-                 nodesize = new_empty_quosure(varying())
-               )
-  )
-  expect_equal(min_n_v_spark$method$fit$args,
-               list(
-                 x = expr(missing_arg()),
-                 formula = expr(missing_arg()),
-                 type = "classification",
-                 min_instances_per_node = new_empty_quosure(varying()),
-                 seed = expr(sample.int(10^5, 1))
-               )
-  )
-
 })
 
 test_that('engine arguments', {
   ranger_imp <- rand_forest(mode = "classification")
   expect_equal(translate(ranger_imp %>% set_engine("ranger", importance = "impurity"))$method$fit$args,
                list(
-                 formula = expr(missing_arg()),
-                 data = expr(missing_arg()),
+                 x = expr(missing_arg()),
+                 y = expr(missing_arg()),
                  case.weights = expr(missing_arg()),
                  importance = new_empty_quosure("impurity"),
                  num.threads = 1,
@@ -239,50 +140,6 @@ test_that('engine arguments', {
                  min_info_gain = new_empty_quosure(2),
                  seed = expr(sample.int(10^5, 1))
                )
-  )
-
-  ranger_samp_frac <- rand_forest(mode = "regression")
-  expect_equal(
-    translate(ranger_samp_frac %>%
-                set_engine("ranger", sample.fraction = varying()))$method$fit$args,
-    list(
-      formula = expr(missing_arg()),
-      data = expr(missing_arg()),
-      case.weights = expr(missing_arg()),
-      sample.fraction = new_empty_quosure(varying()),
-      num.threads = 1,
-      verbose = FALSE,
-      seed = expr(sample.int(10^5, 1))
-    )
-  )
-
-
-  randomForest_votes_v <-
-    rand_forest(mode = "regression")
-  expect_equal(
-    translate(randomForest_votes_v %>%
-                set_engine("randomForest", norm.votes = FALSE, sampsize = varying()))$method$fit$args,
-    list(
-      x = expr(missing_arg()),
-      y = expr(missing_arg()),
-      norm.votes = new_empty_quosure(FALSE),
-      sampsize = new_empty_quosure(varying())
-    )
-  )
-
-  spark_bins_v <-
-    rand_forest(mode = "regression")
-  expect_equal(
-    translate(spark_bins_v %>%
-                set_engine("spark", uid = "id label", max_bins = varying()))$method$fit$args,
-    list(
-      x = expr(missing_arg()),
-      formula = expr(missing_arg()),
-      type = "regression",
-      uid = new_empty_quosure("id label"),
-      max_bins = new_empty_quosure(varying()),
-      seed = expr(sample.int(10^5, 1))
-    )
   )
 
 })